feat(api): capability generation endpoints (GRX-AI-006)
Six capabilities (subject line, body copy, social caption, rewrite,
hashtags, posting time), each a thin capabilities/*.py wrapper
(run(input, provider, model)) around its own prompts/templates.py
builder -- the resolved provider is passed in, never re-resolved inside
the capability, and prompt-building stays separate from the actual
provider.generate() call, so a future multi-agent slice can wrap these
as tools without a rewrite.

POST /ai/generate/{capability} resolves the effective provider (BYO else
platform default else a clean 409), runs the capability, and logs an
ai_generations row (COMPLETE/FAILED) plus a usage_records row on success
-- giving that table's zero-writers gap since Sprint 1 its first real
data. Prompt-injection posture: the system prompt is always fixed
server-authored text plus the account's own brand_voice, customer input
always lands in the user prompt, never the instruction position.

Two real bugs found and fixed via live verification against a real
provider (the user's own Krutrim account, an OpenAI-compatible
third-party API): OpenAIProvider couldn't reach any custom endpoint at
all (hardcoded api.openai.com) -- added an optional SSRF-validated
base_url override; all four adapters silently returned a "successful"
empty generation when a reasoning model hit max_tokens mid-chain-of-
thought -- added an explicit empty-content check to every adapter and
raised every capability's token budget. A third bug (unrelated,
independently found): ai_generations.estimated_cost_usd's unscaled
Numeric column stored float binary-representation artifacts instead of
clean 6-decimal values -- fixed via a new migration.

Live-verified all six capabilities end-to-end against the real Krutrim
API with coherent output, real token counts, and clean costs. This
closes the evidence gap for GRX-AI-003/004/005 too (flipped to DONE
here) -- the successful generation proved the shared adapter/credential-
resolution machinery all four tasks depend on.

+++ apps/api/src/growixa_api/ai/models.py
@@ -58,7 +58,11 @@ class AIGeneration(Base):
     model: Mapped[str] = mapped_column(Text, nullable=False)
     prompt_tokens: Mapped[int | None] = mapped_column(Integer, nullable=True)
     completion_tokens: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    estimated_cost_usd: Mapped[float | None] = mapped_column(Numeric, nullable=True)
+    # Fixed precision/scale (not plain Numeric) -- an unscaled NUMERIC column stores a
+    # Python float's exact binary representation, producing long floating-point
+    # artifacts (e.g. 0.000569999999999999977...) instead of the intended rounded
+    # value. Found live: a real generation's estimated_cost_usd came back this way.
+    estimated_cost_usd: Mapped[float | None] = mapped_column(Numeric(10, 6), nullable=True)
     status: Mapped[str] = mapped_column(Text, nullable=False)
     error_message: Mapped[str | None] = mapped_column(Text, nullable=True)
     # Optional provenance -- which campaign/social_post this generation was made from, set
